taskworker/requirements.txt
python-decouple==3.4
pika==1.2.0
kubernetes==18.20.0
yfinance==0.1.63